+++ internal/config.go
@@ -14,7 +14,7 @@ var config Configuration
 
 // Configuration is a struct to store the script's configuration
 type Configuration struct {
-	RootDomain    string   `yaml:"rootDomain"`
+	RootDomain    string   `yaml:"rootDomain,omitempty"`
 	IPProviders   []string `yaml:"ipProviders"`
 	Notifications struct {
 		Slack struct {
@@ -32,9 +32,11 @@ type Configuration struct {
 	} `yaml:"traefik"`
 	AdditionalRecords []cloudflare.DNSRecord `yaml:"additionalRecords"`
 	Cloudflare        struct {
-		Email    string `yaml:"email"`
-		APIKey   string `yaml:"apiKey"`
-		Defaults struct {
+		Email       string `yaml:"email"`
+		APIKey      string `yaml:"apiKey"`
+		DeleteGrace int    `yaml:"deleteGrace"`
+		ZoneName    string `yaml:"zoneName"`
+		Defaults    struct {
 			Type    string `yaml:"type"`
 			Proxied bool   `yaml:"proxied"`
 			TTL     int    `yaml:"ttl"`
@@ -66,9 +68,15 @@ func GetConfig(configFilePath string) {
 		config.Cloudflare.APIKey = os.Getenv("CLOUDFLARE_APIKEY")
 	}
 
+	// Backwards compatibility
+	if config.RootDomain != "" {
+		config.Cloudflare.ZoneName = config.RootDomain
+		log.Warn("rootDomain is deprecated and will be removed in a future release, use cloudflare.zoneName instead")
+	}
+
 	// Validate config
-	if config.RootDomain == "" {
-		log.Fatal("rootDomain cannot be empty")
+	if config.Cloudflare.ZoneName == "" {
+		log.Fatal("Cloudflare zone name cannot be empty")
 	}
 	if config.Traefik.URL == "" {
 		log.Fatal("Traefik URL cannot be empty")